refactor: update UI
use same code as the API to expand target properties
