+++ apis/core/lib/domain/organization/Organization.ts
@@ -4,13 +4,7 @@ import { Organization } from '@rdfine/schema'
 import { schema } from '@tpluscode/rdf-ns-builders'
 import { DomainError } from '../../errors'
 
-interface CreateIdentifier {
-  cubeIdentifier: string
-  termName?: string | Term
-}
-
 interface ApiOrganization {
-  createIdentifier(params: CreateIdentifier): NamedNode
   updateNamespace(newNamespace: NamedNode | undefined): NamedNode
   updatePublishGraph(publishGraph: Term | undefined): void
 }
@@ -23,32 +17,6 @@ declare module '@rdfine/schema' {
 
 export default function Mixin<Base extends Constructor<Omit<Organization, keyof ApiOrganization>>>(Resource: Base) {
   class Organization extends Resource implements ApiOrganization {
-    createIdentifier({ cubeIdentifier, termName }: CreateIdentifier): NamedNode {
-      const namespace = this.namespace.value.match(/[/#]$/) ? this.namespace.value : `${this.namespace.value}/`
-      let base = namespace + cubeIdentifier
-
-      if (!termName) {
-        return this.pointer.namedNode(base).term
-      }
-
-      if (!base.match(/[/#]$/)) {
-        base += '/'
-      }
-
-      if (typeof termName === 'string') {
-        return this.pointer.namedNode(base + termName).term
-      }
-
-      if (termName.termType === 'Literal') {
-        return this.pointer.namedNode(base + termName.value).term
-      }
-      if (termName.termType === 'NamedNode') {
-        return termName
-      }
-
-      throw new Error(`Unexpected identifier template type ${termName.termType}`)
-    }
-
     updateNamespace(newNamespace: NamedNode | undefined) {
       if (!newNamespace) {
         throw new DomainError('Namespace cannot be empty')
