+++ ui/src/store/modules/project.ts
@@ -24,8 +24,14 @@ import {
 } from '../serializers'
 import { Term } from 'rdf-js'
 import { RdfResource } from 'alcaeus'
+import type { Organization } from '@rdfine/schema'
+
+export interface CreateIdentifier {
+  (term: string): string
+}
 
 export interface ProjectState {
+  createIdentifier: null | CreateIdentifier,
   project: null | Project,
   csvMapping: null | CsvMapping,
   sourcesCollection: null | SourcesCollection,
@@ -38,6 +44,7 @@ export interface ProjectState {
 }
 
 const initialState = {
+  createIdentifier: null,
   project: null,
   csvMapping: null,
   sourcesCollection: null,
@@ -241,8 +248,19 @@ const actions: ActionTree<ProjectState, RootState> = {
 }
 
 const mutations: MutationTree<ProjectState> = {
-  storeProject (state, project) {
+  storeProject (state, project: Project) {
     state.project = Object.freeze(project)
+    if (project) {
+      const { cubeIdentifier } = project
+      state.createIdentifier = (termName: string) => {
+        return (project.maintainer as Organization).createIdentifier({
+          termName,
+          cubeIdentifier,
+        }).value
+      }
+    } else {
+      state.createIdentifier = null
+    }
   },
 
   storeCSVMapping (state, mapping) {
