+++ ui/src/views/CubeProjectEdit.vue
@@ -38,7 +38,6 @@ import { Project } from '@cube-creator/model'
 import HydraOperationForm from '@/components/HydraOperationForm.vue'
 import { GraphPointer } from 'clownface'
 import { Shape } from '@rdfine/shacl'
-import { cc } from '@cube-creator/core/namespace'
 import { api } from '@/api'
 import { APIErrorValidation, ErrorDetails } from '@/api/errors'
 
@@ -62,8 +61,7 @@ export default class CubeProjectEditView extends Vue {
       this.shape = await api.fetchOperationShape(operation)
     }
 
-    const namespace = this.project.csvMapping?.namespace ?? ''
-    this.resource = Object.freeze(this.project.pointer.addOut(cc.namespace, namespace))
+    this.resource = Object.freeze(this.project.pointer)
   }
 
   async updateProject (): Promise<void> {
